add DataForDc in StorageWiggleData; rename KeyBackedConfig to KeyBackedStruct
